NST-36 Update inputs to read from config
